Fix: GPv3: Draw Tool generates invalid vertex colors
When the drawing contains a stroke that was drawn with vertex colors,
drawing a new stroke with just the material color can lead to invalid
vertex colors.

This was because the draw tool did not write to the vertex color attribute
even if it existed. This meant that when the curves geometry got
resized to be larger, the attribute would contain invalid values for
the new points.

The fix makes sure to write empty vertex colors (black, fully transparent)
to the attribute if the attribute already exists or the vertex color
is being used.