UI: Use Official Blender Logo on Top Bar
Our icon sources now include two versions of the Blender logo, the
official one and one that is modified to align better to our smallest
pixel grid. But with our recent change to SVG icons, and alignment
tweaks to the official version I think we can use it in more cases.
This PR uses it on the Top Bar.