Anim: add concept of 'active slot'
Add the concept of the 'active slot' within an Action. This is not yet
used in the code; it is intended to integrate into the Action editor and
thus the animation filtering code. I'd rather keep the PRs that touch
filtering code changes to a minimum, and so I'd like to commit this
separately.