File Browser: add preferences to View menu
The file browser has very few display settings, so they should be easily discoverable and not
spread across different places.

While it is a bit of an anti-pattern to show global settings in an editor, I feel like for the file browser,
most users would expect the settings there to be global anyway.