WIP: Anim: Copy Global Transform for multiple bones & objects
Add the ability for the Copy Global Transform operator to copy/paste
transforms of multiple bones & objects. This will then be done by name,
rather than just copying from/to the active bone/object.

It's starting to work, but care needs to be taken in the order in which
the matrices are applied (parents before children) to get the right
result.