Fix incorrect index when interpolating between different curve amounts.
The `selection` is a bit difficult to interpret: It contains indices in
the _destination_ curve (the interpolated result), not any of the input
geometries. The `from_curve_indices`/`to_curve_indices` slices have the
same size as the selection. When iterating over the selection the first
index argument is the _destination_ index (selection contains
destination indices), while the second index is the position in the
to/from_curve_indices.