Fix #124133: Compute actual transform between domains outside glsl
Move the "offset" from the glsl shader and put it into the matrix
that is passed to it. This makes the matrix be the actual transformation
between the output and input domains (with 0,0 in the lower-left corner
of each), which should make it easier to analyze what it is doing.