UI: Added image size control
This feature is useful for users who want to maintain a consistent
scale for their reference images without having to manually
take action to scale/edit the size of their images. Additionally,
this pairs well with the previously added multiple image importing,
since now users can set all of their imported reference images to a
predetermined size during the import process.

Added a float property to OBJECT_OT_empty_image_add containing the
size, and set the empty_drawsize of each new object to be
equal to the float property.