UI: Color Icons
This PR allows SVG UI icons to be shown in full color by changing them
from DEF_ICON to DEF_ICON_COLOR.  This also adds support for an
optional callback that can be used to alter the SVG source before
rasterization. That callback is unused here but will be part of
later PRs.