Fix #124298: Cryptomatte node makes UI very slow
The Cryptomatte node makes the UI very slow in Image mode. That's
because the node updates its available Cryptomatte layers on every
redraw, which is a blocking operation that is not very fast.

To fix this, we only update the layers when engine layers change for
the Scene mode, and when image settings change for Image mode.

Cryptomatte layer initialization required the active scene to update the
image user frame, which is information that is not available in the
updatefunc of the node type, so the scene argument was removed from the
call chain, since updating the image user is already done through other
mechanisms and is redundant to do at redraw time. And the frame number
of the image user is ignored at compositing time.