Fix sampling calls in CPU compositor
These are adjusted to match the locations used by the GPU compositor,
which I believe is correct. Mostly this means adding .5 to the output
point before the transform, and subtracting .5 to get the input point.

I also fixed the Scale Absolute bounding box to produce the requested
size. Rounding errors were causing it to produce larger sizes.

COM_Translate is unchanged as the position is correct. However it is
putting transparent rows of pixel on the bottom and left edges of the
output rather than the correct value, I was unable to figure out the
cause of this.