Fix #124133: Put the .5 shift into RealizeOnDomainOperation
The half pixel shift removed by the previous patch is re-introduced
in a much more restricted domain. It's purpose is to not produce
unexpected filtering if an odd-sized image is mixed with an
even-sized one.

This also reverses the shift direction depending on which image is
even. I beleive this will prevent these corrections from accumulating
into multiple pixel shifts.