CMake: Add buildbot configuration for all platforms
It allows to more easily override cmake options which we want to
be enabled on the buildbot.

In a way it was possible to do via the pipeline config, but this
change provides a more natural way for Blender developers to set
CMake configuration which is to be enabled on buildbot, but not
on local developer machines.