Fix #124842: Prevent rounding errors from introducing black pixels in Translate
Since translate only reads pixels inside the buffer, it is safe to treat Clip as
Extend.

IMHO the actual bug is in COM_MemoryBuffer. wrap_pixel() must treat Clip as Extend,
and code rearranged so calls to it are done after the clipping tests. For Nearest
sampling the sample must be less than -0.5 to produce black. For Bilinear sampling
samples between -1 and 0 must linearly interpolate between black and the edge pixel.