Fix #124589: VSE meta/scene strip contents have misleading horizontal margins
To account for rounded corners, the meta/scene strip contents were
drawn with a horizontal margin on both strip sides. However this is
confusing and misleading, since it can look like the contents have
a frame gap between content start and parent strip start.

Instead, add vertical margins which do not have a possibility of
confusion with frame gaps.