Modeling: remove unused Knife cut operator
This got superseeded by the "super knife" in 421823e34e5d 14 years ago
it seems, it cannot be executed atm. (and even if reactivated, it
crashes in multiple places).

Stumbled over this while checking on usages of
#BM_custom_loop_normals_to_vector_layer /
#BM_custom_loop_normals_from_vector_layer (to add this to more tools),
and apparently this outdated piece of code got an update in 93c8955a722b
:) -- work on adding this to our "real" Knife is upcoming in another
PR...