Fix #65926: Compositor executes redundantly on undo
The interactive compositor executes redundantly on any undo. Which is
also problematic for operators when adjusting their settings. To fix
this, we no longer execute the compositor on undo, but we execute on
changes to the compositor node tree, which should cover undo on only the
resources used by the compositor.