Fix (unreported) 4.2: Copy of material ID fails.
Copy/paste code of Material ID will generate an empty pastebuffer
blendfile, due to changes in how IDs are checked for being used or not
in general writefile code.

Yet another failure due to invalid design of current partial write code
(this is being addressed in main/4.3 code), and lack of unittests (will
be addresed in main asap as well).