Revert #124938: Allow Back Button When File Browser Loaded in Workspace
This reverts #124938 & #124827, leaving File Browser as it was before,
without any extra ED_fileselect_set_params_from_userdef to ensure that
the previous_dir is populated with current directory. I have not found
a place to ensure this history is correct (when File Browser is part of
the layout) that does not cause unintended consequences. Will mark the
original bug report, #124771, as unresolved.