UI: Icon Hover Brightness for Properties Items
Properties category icons with new SVG icons are slightly brighter than
before, and therefore not seem to brighten on hover. This PR makes
them match exactly. Was assigning an alpha instead of multiplying. And
and earlier refactor for outlines made these icons not show initially
dimmer unless there is also an outline.