Fix #124842: Fix clipping and Nearest sampling in COM_MemoryBuffer::read/read_elem_sampled
This uses the math_interp.cc code with Extend (or Repeat if possible)
and then multiplies by a correctly-antialiased clipping rectangle. I had
to reverse which function calls the other so that the two extend parameters
were accessible.

The Nearest sampling also had an off-by-.5 error. This is probably a mistake
in math_interp but I changed the calling code instead.