Anim: forward compatibility for F-Curves in layered Actions
When writing a layered Action to disk, take the F-Curves from the first
keyframe strip and write that as `action.curves` as well. This will make
older Blender versions see those curves and load them properly.

Only the curves for the first slot are written this way. This means that
any legacy Action that was converted to a layered Action will be loaded
again properly by older Blender versions. Of course this is limited to a
single layer, single strip, and single slot -- once the newer features
are used, older versions of Blender will not be able to see this extra
data.

When an Action contains multiple slots, so with animation for multiple
distinct objects, the forward compatibility becomes a bit iffy. Older
versions of Blender will just see a legacy Action, with its legacy
semantics, and thus all objects that use that Action will receive the
exact same animation data. I don't think there's a way around this.
(Unless we start breaking up Actions into an Action per slot, alter the
assignments, and then store metadata so that modern Blenders can
reassemble them. I do not think this is a good idea.)

Ref: #124714