Sculpt/paint: Write brush types to brush asset metadata
This way we can query the brush types, e.g. to filter out eraser brushes
in the brush asset selector while the eraser tool is selected.

Technically a brush may have different brush types depending on the
mode. So we store a type for all supported modes.

Ideally the name of the custom metadata property would match the name of
the properties in RNA. However these should be renamed to avoid the term
"tool", see #124201, and we currently don't do versioning of asset
metadata. So this uses what I expect the name will be. Could create a
separate PR to duplicate the RNA properties with the new name.