Fix #125063: Anim: layered Action data is still loaded in 4.2-release
Add `#ifdef WITH_ANIM_BAKLAVA` to the blend file reading/writing code,
so that the Action layers & slots are ignored.

This ensures that any loaded Action is just treated as 'legacy' (which
is the only kind of Action Blender 4.2 should have to deal with), which
will also deal with the forward compatible data written by !125065.